feat: Implement class-based dark mode and ensure form input contrast
