+++ app/metadata.tsx
@@ -1,0 +1,10 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "HostelHub",
+  description: "Find the perfect hostel for your stay",
+  icons: {
+    icon: '/assets/icons/hostelhubFavicon.svg',
+    shortcut: '/assets/icons/hostelhubFavicon.svg',
+  },
+};
